refactor(paths): a path predicate belongs with the path rules
